requirements.txt
# Core dependencies
langchain>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ipywidgets>=8.0.0
jupyter>=1.0.0
notebook>=7.0.0

# LLM providers
ollama>=0.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
plotly>=5.18.0
pillow>=10.0.0

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# LLM and Vector Store
openai>=1.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# Utilities
tqdm>=4.66.1
loguru>=0.7.2
duckduckgo-search>=3.9.3
beautifulsoup4>=4.9.3
aiohttp>=3.8.0

# Testing
pytest-asyncio>=0.21.1

# Monitoring
psutil>=5.9.0
redis>=5.0.0
aiohttp>=3.9.0

# Jupyter
ipykernel>=6.0.0
jupyterlab>=4.0.0

# Visualization
graphviz>=0.20.1 